requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-gis==0.16
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
googlemaps==4.4.2
idna==2.10
numpy==1.19.1
Pillow==7.2.0
psycopg2-binary==2.8.5
pygdal==2.4.2.5
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
